api/requirements.txt
fastapi
pydantic
pymongo
python-dotenv
uvicorn
motor
typing